move index.ts
