+++ lib/Providers.d.ts
@@ -1,5 +1,5 @@
-import IWalletProvider from "../web3bch-providers/IWalletProvider";
-import INetworkProvider from "../web3bch-providers/INetworkProvider";
+import IWalletProvider from "./web3bch-providers/IWalletProvider";
+import INetworkProvider from "./web3bch-providers/INetworkProvider";
 export default class Providers {
     networkProvider?: INetworkProvider | undefined;
     walletProvider?: IWalletProvider | undefined;
