+++ lib/index.js
@@ -1,7 +1,7 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const Providers_1 = require("./Providers");
-const Wallet_1 = require("../web3bch-wallet/Wallet");
+const Wallet_1 = require("./web3bch-wallet/Wallet");
 class Web3bch {
     constructor() {
         /**
