+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "description": "",
   "scripts": {
-    "build": "tsc",
+    "build": "rm -rf lib && tsc",
     "docs": "npm run docs:build && npm run docs:touch-nojekyll",
     "docs:build": "typedoc --out docs src --plugin typedoc-plugin-example-tag",
     "docs:touch-nojekyll": "touch docs/.nojekyll",
