+++ src/Providers.ts
@@ -1,5 +1,5 @@
-import IWalletProvider from "../web3bch-providers/IWalletProvider"
-import INetworkProvider from "../web3bch-providers/INetworkProvider"
+import IWalletProvider from "./web3bch-providers/IWalletProvider"
+import INetworkProvider from "./web3bch-providers/INetworkProvider"
 
 export default class Providers {
   constructor(public networkProvider?: INetworkProvider, public walletProvider?: IWalletProvider) {}
